Use Context for ArticleRenderer
