+++ app/blog/renderer/ArticleRenderer.tsx
@@ -3,19 +3,19 @@
 import React, {useContext} from "react";
 import ReactMarkdown from "react-markdown";
 import RendererContext from "./RendererContext";
+import {MathJaxWrapper} from "../../../components/utils/MathJaxWrapper";
 
 export default function ArticleRendererFromContext ({toRender}: {
   toRender: string,
 }) {
-  const options = useContext(RendererContext);
-  const {debugStr} = options;
-  delete options.debugStr;
+  const context = useContext(RendererContext);
   return (
     <>
-      {debugStr ?? ''}
-      <ReactMarkdown {...options as any}>
-        {toRender}
-      </ReactMarkdown>
+      <MathJaxWrapper>
+        <ReactMarkdown {...context.markdown.options as any}>
+          {toRender}
+        </ReactMarkdown>
+      </MathJaxWrapper>
     </>
   )
 }
