+++ app/blog/renderer/RendererContext.tsx
@@ -3,12 +3,16 @@
 import {createContext} from "react";
 import {ReactMarkdownOptions} from "react-markdown/lib/react-markdown";
 
-const RendererContext = (
-  createContext<Partial<ReactMarkdownOptions & {
-    debugStr: string
-  }>>({
-    debugStr: 'not initialized',
-  })
-);
+type ContextProp = {
+  markdown: {
+    options: Partial<ReactMarkdownOptions>
+  }
+}
+
+const RendererContext = createContext<ContextProp>({
+  markdown: {
+    options: {}
+  }
+});
 
 export default RendererContext;
